refactor(models): move CollectableElement to ElementCollection, re-home ARElement/PackageableElement re-exports, flatten RolesAndRights per spec (Rule 0007 A)
