+++ src/armodel/models/M2/AUTOSARTemplates/CommonStructure/__init__.py
@@ -2,6 +2,11 @@
 CommonStructure module re-exports classes from submodules for backward compatibility.
 """
 
+# Import AbstractBlueprintStructure (a light module) first: AbstractStructure imports it,
+# and its parent-package initialization must not re-enter AbstractStructure while the
+# latter is only partially initialized (see Task 15 bootstrap-cycle fix).
+from armodel.models.M2.AUTOSARTemplates.CommonStructure.StandardizationTemplate.AbstractBlueprintStructure import AtpBlueprintable  # noqa: F401
+
 from armodel.models.M2.AUTOSARTemplates.CommonStructure.Constants import (
     ValueSpecification,
     CompositeValueSpecification,
