+++ src/armodel/models/M2/AUTOSARTemplates/DiagnosticExtract/CommonDiagnostics.py
@@ -1,5 +1,5 @@
 from abc import ABC
-from armodel.models.M2.AUTOSARTemplates.GenericStructure.GeneralTemplateClasses.Identifiable import (
+from armodel.models.M2.AUTOSARTemplates.GenericStructure.GeneralTemplateClasses.ARPackage import (
     ARElement,
 )
