+++ src/armodel/models/M2/AUTOSARTemplates/GenericStructure/GeneralTemplateClasses/ARPackage.py
@@ -7,6 +7,25 @@ components, interfaces, data types, and other packages.
 """
 
 from typing import Dict, List, Optional, Any
+from typing import TYPE_CHECKING
+
+if TYPE_CHECKING:
+    from armodel.models.M2.AUTOSARTemplates.GenericStructure.GeneralTemplateClasses.ElementCollection import Collection
+
+from armodel.models.M2.AUTOSARTemplates.GenericStructure.GeneralTemplateClasses.Identifiable import (  # noqa: F401
+    ARElement,
+    Identifiable,
+    PackageableElement,
+    Referrable,
+)
+
+# Initialize the CommonStructure package before any import that transitively touches
+# AbstractStructure: AbstractStructure's own import of AbstractBlueprintStructure requires
+# the CommonStructure package to be present in sys.modules (Task 15 bootstrap-cycle fix).
+import armodel.models.M2.AUTOSARTemplates.CommonStructure  # noqa: F401,E402
+
+from armodel.models.M2.AUTOSARTemplates.GenericStructure.GeneralTemplateClasses.ElementCollection import CollectableElement  # noqa: E402
+
 
 from armodel.models.M2.MSR.DataDictionary.SystemConstant import SwSystemconst
 from armodel.models.M2.MSR.Documentation.TextModel.BlockElements import DocumentationBlock
@@ -38,8 +57,8 @@ from armodel.models.M2.MSR.DataDictionary.RecordLayout import SwRecordLayout
 from armodel.models.M2.AUTOSARTemplates.BswModuleTemplate.BswImplementation import BswImplementation
 from armodel.models.M2.AUTOSARTemplates.BswModuleTemplate.BswOverview import BswModuleDescription
 from armodel.models.M2.AUTOSARTemplates.BswModuleTemplate.BswInterfaces import BswModuleEntry
-from armodel.models.M2.AUTOSARTemplates.CommonStructure import ConstantSpecification
 from armodel.models.M2.AUTOSARTemplates.CommonStructure.ImplementationDataTypes import ImplementationDataType
+from armodel.models.M2.AUTOSARTemplates.CommonStructure import ConstantSpecification
 from armodel.models.M2.AUTOSARTemplates.CommonStructure.Implementation import Implementation
 from armodel.models.M2.AUTOSARTemplates.CommonStructure.FlatMap import FlatMap
 from armodel.models.M2.AUTOSARTemplates.CommonStructure.McGroups import McGroup
@@ -56,8 +75,6 @@ from armodel.models.M2.AUTOSARTemplates.EcuResourceTemplate import HwElement
 from armodel.models.M2.AUTOSARTemplates.ECUCParameterDefTemplate import EcucDefinitionCollection, EcucDestinationUriDefSet, EcucModuleDef
 from armodel.models.M2.AUTOSARTemplates.EcuResourceTemplate.HwElementCategory import HwCategory, HwType
 from armodel.models.M2.AUTOSARTemplates.GenericStructure.GeneralTemplateClasses.ArObject import ARObject
-from armodel.models.M2.AUTOSARTemplates.GenericStructure.GeneralTemplateClasses.ElementCollection import Collection
-from armodel.models.M2.AUTOSARTemplates.GenericStructure.GeneralTemplateClasses.Identifiable import CollectableElement, Referrable
 from armodel.models.M2.AUTOSARTemplates.GenericStructure.GeneralTemplateClasses.PrimitiveTypes import Boolean, Identifier, RefType, ReferrableSubtypesEnum
 from armodel.models.M2.AUTOSARTemplates.GenericStructure.LifeCycles import LifeCycleInfoSet
 from armodel.models.M2.AUTOSARTemplates.SWComponentTemplate.Composition import CompositionSwComponentType
@@ -1473,7 +1490,9 @@ class ARPackage(CollectableElement):
             self.addElement(props)
         return self.getElement(short_name, E2EProfileCompatibilityProps)
 
-    def createCollection(self, short_name: str) -> Collection:
+    def createCollection(self, short_name: str) -> "Collection":
+        from armodel.models.M2.AUTOSARTemplates.GenericStructure.GeneralTemplateClasses.ElementCollection import Collection
+
         if not self.IsElementExists(short_name, Collection):
             collection = Collection(self, short_name)
             self.addElement(collection)
@@ -1708,7 +1727,9 @@ class ARPackage(CollectableElement):
     def getDataTransformationSets(self) -> List[DataTransformationSet]:
         return list(sorted(filter(lambda a: isinstance(a, DataTransformationSet), self.elements), key=lambda a: a.short_name))
 
-    def getCollections(self) -> List[Collection]:
+    def getCollections(self) -> List["Collection"]:
+        from armodel.models.M2.AUTOSARTemplates.GenericStructure.GeneralTemplateClasses.ElementCollection import Collection
+
         return list(sorted(filter(lambda a: isinstance(a, Collection), self.elements), key=lambda a: a.short_name))
 
     def getKeywordSets(self) -> List[KeywordSet]:
