+++ src/armodel/models/M2/AUTOSARTemplates/GenericStructure/GeneralTemplateClasses/ElementCollection.py
@@ -3,12 +3,133 @@ This module contains classes for representing AUTOSAR element collections
 in the GenericStructure module.
 """
 
-from typing import List, Optional
+from abc import ABC
+from typing import Dict, List, Optional
 
+from armodel.models.M2.AUTOSARTemplates.GenericStructure.GeneralTemplateClasses.Identifiable import Referrable
+from armodel.models.M2.AUTOSARTemplates.GenericStructure.GeneralTemplateClasses.ArObject import ARObject
 from armodel.models.M2.AUTOSARTemplates.GenericStructure.GeneralTemplateClasses.PrimitiveTypes import Identifier, NameToken, RefType
 from armodel.models.M2.AUTOSARTemplates.GenericStructure.GeneralTemplateClasses.AnyInstanceRef import AnyInstanceRef
-from armodel.models.M2.AUTOSARTemplates.GenericStructure.GeneralTemplateClasses.Identifiable import ARElement
 from armodel.models.M2.AUTOSARTemplates.GenericStructure.GeneralTemplateClasses.Enumerations import AutoCollectEnum
+
+
+class CollectableElement(ARObject, ABC):
+    """
+    Abstract class for elements that can collect other referrable elements.
+    This class provides functionality for managing collections of elements with lookup capabilities.
+    """
+
+    # CollectableElement method parity checklist:
+    # [ ] __init__                     [x] impl  [ ] docstring  [ ] test
+    # [x] getTotalElement              [x] impl  [x] docstring  [x] test
+    # [x] removeElement                [x] impl  [x] docstring  [x] test
+    # [ ] getElements                  [x] impl  [x] docstring  [ ] test
+    # [ ] addElement                   [x] impl  [x] docstring  [ ] test
+    # [ ] getElement                   [x] impl  [x] docstring  [ ] test
+    # [x] IsElementExists              [x] impl  [x] docstring  [x] test
+
+    def __init__(self):
+        if type(self) is CollectableElement:
+            raise TypeError("CollectableElement is an abstract class.")
+
+        super().__init__()
+
+        self.elements: List[Referrable] = []
+        self.element_mappings: Dict[str, List[Referrable]] = {}
+
+    def getTotalElement(self) -> int:
+        """
+        Gets the total number of elements in this collection.
+
+        Returns:
+            The count of elements in the collection
+        """
+        # return len(list(filter(lambda a: not isinstance(a, ARPackage) , self.elements)))
+        return len(self.elements)
+
+    def removeElement(self, short_name: str, type=None):
+        """
+        Removes an element from this collection.
+
+        Args:
+            short_name: The short name of the element to remove
+            type: The type of element to remove (optional)
+        """
+        if short_name not in self.element_mappings:
+            raise KeyError("Invalid key <%s> for removing element" % short_name)
+        if type is None:
+            item = self.element_mappings[short_name][0]
+        else:
+            item = next(filter(lambda a: isinstance(a, type), self.element_mappings[short_name]))
+        if item is not None:
+            self.elements.remove(item)
+            self.element_mappings[short_name].remove(item)
+
+    def getElements(self) -> List[Referrable]:
+        """
+        Gets the list of elements in this collection.
+
+        Returns:
+            List of Referrable instances
+        """
+        return self.elements
+
+    def addElement(self, element: Referrable):
+        """
+        Adds an element to this collection.
+
+        Args:
+            element: The element to add
+
+        Returns:
+            self for method chaining
+        """
+        short_name = element.getShortName()
+        if not self.IsElementExists(short_name, type(element)):
+            self.elements.append(element)
+            if short_name not in self.element_mappings:
+                self.element_mappings[short_name] = []
+            self.element_mappings[short_name].append(element)
+
+    def getElement(self, short_name: str, type=None) -> Optional[Referrable]:
+        """
+        Gets an element from this collection by short name and type.
+
+        Args:
+            short_name: The short name of the element to find
+            type: The type of element to find (optional)
+
+        Returns:
+            The found Referrable instance, or None if not found
+        """
+        if short_name not in self.element_mappings:
+            return None
+        if type is not None:
+            result = list(filter(lambda a: isinstance(a, type), self.element_mappings[short_name]))
+            if len(result) == 0:
+                return None
+            return result[0]
+        return self.element_mappings[short_name][0]
+
+    def IsElementExists(self, short_name: str, type=None) -> bool:
+        """
+        Checks if an element with the specified short name and type exists in this collection.
+
+        Args:
+            short_name: The short name of the element to check
+            type: The type of element to check (optional)
+
+        Returns:
+            True if the element exists, False otherwise
+        """
+        if type is None:
+            return short_name in self.element_mappings
+        if short_name in self.element_mappings:
+            return any(isinstance(a, type) for a in self.element_mappings[short_name])
+        return False
+
+
+from armodel.models.M2.AUTOSARTemplates.GenericStructure.GeneralTemplateClasses.ARPackage import ARElement  # noqa: E402
 
 
 class Collection(ARElement):
