--- src/armodel/models/M2/AUTOSARTemplates/GenericStructure/RolesAndRights/__init__.py
@@ -1,7 +1,0 @@
-"""
-RolesAndRights module for AUTOSAR M2 models.
-"""
-
-from armodel.models.M2.AUTOSARTemplates.GenericStructure.RolesAndRights.AtpDefinition import AtpDefinition
-
-__all__ = ["AtpDefinition"]
