+++ src/armodel/models/__init__.py
@@ -187,7 +187,7 @@ from armodel.models.M2.AUTOSARTemplates.GenericStructure.GeneralTemplateClasses.
 from armodel.models.M2.AUTOSARTemplates.GenericStructure.GeneralTemplateClasses.PrimitiveTypes import *  # noqa: F403
 from armodel.models.M2.AUTOSARTemplates.GenericStructure.GeneralTemplateClasses.TagWithOptionalValue import *  # noqa: F403
 from armodel.models.M2.AUTOSARTemplates.GenericStructure.LifeCycles import *  # noqa: F403
-from armodel.models.M2.AUTOSARTemplates.GenericStructure.RolesAndRights.AtpDefinition import *  # noqa: F403
+from armodel.models.M2.AUTOSARTemplates.GenericStructure.RolesAndRights import *  # noqa: F403
 from armodel.models.M2.AUTOSARTemplates.GenericStructure.DocumentationOnM1 import *  # noqa: F403
 
 # Additional SWComponentTemplate imports
