+++ tests/test_armodel/models/M2/AUTOSARTemplates/GenericStructure/GeneralTemplateClasses/test_Identifiable.py
@@ -5,12 +5,9 @@ in the AUTOSAR GenericStructure module.
 
 from armodel.models.M2.AUTOSARTemplates.AutosarTopLevelStructure import AUTOSAR
 from armodel.models.M2.AUTOSARTemplates.GenericStructure.GeneralTemplateClasses.Identifiable import (
-    ARElement,
-    CollectableElement,
     Describable,
     Identifiable,
     MultilanguageReferrable,
-    PackageableElement,
     Referrable,
     ShortNameFragment,
 )
@@ -238,385 +235,6 @@ class TestMultilanguageReferrable:
         assert obj.getLongName() is long_name
 
 
-class TestCollectableElement:
-    """
-    Test class for CollectableElement functionality.
-    """
-
-    def test_abstract_initialization(self):
-        """
-        Test that CollectableElement cannot be instantiated directly (abstract class).
-        """
-        try:
-            _obj = CollectableElement()
-            assert False, "CollectableElement should not be instantiable"
-        except TypeError:
-            pass  # Expected behavior
-
-    def test_get_total_element(self):
-        """
-        Test getTotalElement method.
-        """
-
-        class ConcreteCollectableElement(CollectableElement):
-            def __init__(self):
-                super().__init__()
-
-        obj = ConcreteCollectableElement()
-        assert obj.getTotalElement() == 0
-
-        # Add an element
-        class MockReferrable:
-            def __init__(self, short_name):
-                self.short_name = short_name
-
-            def getShortName(self):
-                return self.short_name
-
-        mock_element = MockReferrable("TestElement")
-        obj.addElement(mock_element)
-        assert obj.getTotalElement() == 1
-
-    def test_add_element_and_get_elements(self):
-        """
-        Test addElement and getElements methods.
-        """
-
-        class ConcreteCollectableElement(CollectableElement):
-            def __init__(self):
-                super().__init__()
-
-        obj = ConcreteCollectableElement()
-
-        # Initially should be empty
-        assert obj.getElements() == []
-
-        # Add an element
-        class MockReferrable:
-            def __init__(self, short_name):
-                self.short_name = short_name
-
-            def getShortName(self):
-                return self.short_name
-
-        mock_element = MockReferrable("TestElement")
-        obj.addElement(mock_element)
-
-        elements = obj.getElements()
-        assert len(elements) == 1
-        assert elements[0] == mock_element
-
-    def test_get_element_with_type(self):
-        """
-        Test getElement method with type parameter to cover missing lines.
-        """
-
-        class ConcreteCollectableElement(CollectableElement):
-            def __init__(self):
-                super().__init__()
-
-        obj = ConcreteCollectableElement()
-
-        # Add an element - use a proper Referrable implementation
-        class ConcreteReferrable(Referrable):
-            def __init__(self, parent, short_name):
-                super().__init__(parent, short_name)
-
-        parent = AUTOSAR.getInstance()
-        ar_root = parent.createARPackage("AUTOSAR")
-        mock_element = ConcreteReferrable(ar_root, "TestElement")
-        obj.addElement(mock_element)
-
-        # Test getting element with specific type
-        result = obj.getElement("TestElement", type=ConcreteReferrable)
-        assert result == mock_element
-
-        # Test getting element with wrong type (should return None)
-        result = obj.getElement("TestElement", type=str)  # Wrong type
-        assert result is None
-
-        # Test getting non-existent element with type
-        result = obj.getElement("NonExistent", type=ConcreteReferrable)
-        assert result is None
-
-    def test_get_element_no_match_for_type(self):
-        """
-        Test getElement method when no elements match the specified type to cover missing line.
-        """
-
-        class ConcreteCollectableElement(CollectableElement):
-            def __init__(self):
-                super().__init__()
-
-        obj = ConcreteCollectableElement()
-
-        # Add an element of one type
-        class ConcreteReferrable(Referrable):
-            def __init__(self, parent, short_name):
-                super().__init__(parent, short_name)
-
-        parent = AUTOSAR.getInstance()
-        ar_root = parent.createARPackage("AUTOSAR")
-        mock_element = ConcreteReferrable(ar_root, "TestElement")
-        obj.addElement(mock_element)
-
-        # Try to get element with different type (should return None)
-        result = obj.getElement("TestElement", type=str)  # Wrong type
-        assert result is None
-
-    def test_get_element_no_match_for_type_manually_added(self):
-        """
-        Test getElement method with manually added elements to ensure filter returns empty list.
-        """
-
-        class ConcreteCollectableElement(CollectableElement):
-            def __init__(self):
-                super().__init__()
-
-        obj = ConcreteCollectableElement()
-
-        # Manually add elements to test the filter with no matches
-        class TypeA:
-            def getShortName(self):
-                return "TestElement"
-
-        class TypeB:
-            def getShortName(self):
-                return "TestElement"
-
-        # Add elements with same name but different types to element_mappings
-        obj.element_mappings["TestElement"] = [TypeA()]
-        obj.elements = [TypeA()]
-
-        # Try to get element with typeB (should return None, triggering the len(result) == 0 path)
-        result = obj.getElement("TestElement", type=TypeB)
-        assert result is None
-
-    def test_is_element_exists_with_type(self):
-        """
-        Test IsElementExists method with type parameter to cover missing lines.
-        """
-
-        class ConcreteCollectableElement(CollectableElement):
-            def __init__(self):
-                super().__init__()
-
-        obj = ConcreteCollectableElement()
-
-        # Initially should return False
-        assert obj.IsElementExists("NonExistent", type=str) is False
-
-        # Add an element
-        class ConcreteReferrable(Referrable):
-            def __init__(self, parent, short_name):
-                super().__init__(parent, short_name)
-
-        parent = AUTOSAR.getInstance()
-        ar_root = parent.createARPackage("AUTOSAR")
-        mock_element = ConcreteReferrable(ar_root, "TestElement")
-        obj.addElement(mock_element)
-
-        # Should return True for correct type
-        assert obj.IsElementExists("TestElement", type=ConcreteReferrable) is True
-
-        # Should return False for incorrect type
-        assert obj.IsElementExists("TestElement", type=str) is False
-
-    def test_remove_element_with_type_param(self):
-        """
-        Test removeElement method with type parameter to cover missing lines.
-        """
-
-        class ConcreteCollectableElement(CollectableElement):
-            def __init__(self):
-                super().__init__()
-
-        obj = ConcreteCollectableElement()
-
-        # Add an element
-        class ConcreteReferrable(Referrable):
-            def __init__(self, parent, short_name):
-                super().__init__(parent, short_name)
-
-        parent = AUTOSAR.getInstance()
-        ar_root = parent.createARPackage("AUTOSAR")
-        mock_element = ConcreteReferrable(ar_root, "TestElement")
-        obj.addElement(mock_element)
-
-        # Call removeElement with type specified to exercise the code path
-        # First, add another element with the same name (this would typically not be done in practice
-        # but is needed to test the type filtering code)
-        # Actually, addElement doesn't allow duplicate names for same type by default
-        # so let's just call the method to ensure the type path is covered
-        original_total = obj.getTotalElement()
-        try:
-            obj.removeElement("TestElement", type=ConcreteReferrable)
-            # If successful, one element should be removed
-            assert obj.getTotalElement() == original_total - 1
-        except StopIteration:
-            # This can happen if type filtering doesn't find an element, which is also a code path
-            pass  # This is also a valid execution path
-
-    def test_remove_element_keyerror_path(self):
-        """
-        Test the KeyError path in removeElement method.
-        """
-
-        class ConcreteCollectableElement(CollectableElement):
-            def __init__(self):
-                super().__init__()
-
-        obj = ConcreteCollectableElement()
-
-        # Try to remove non-existent element to trigger KeyError
-        try:
-            obj.removeElement("NonExistentElement")
-            assert False, "Should have raised KeyError"
-        except KeyError:
-            pass  # Expected behavior
-
-    def test_get_element_default_type(self):
-        """
-        Test getElement method with default type=None to cover line 201.
-        """
-
-        class ConcreteCollectableElement(CollectableElement):
-            def __init__(self):
-                super().__init__()
-
-        obj = ConcreteCollectableElement()
-
-        # Add an element
-        class ConcreteReferrable(Referrable):
-            def __init__(self, parent, short_name):
-                super().__init__(parent, short_name)
-
-        parent = AUTOSAR.getInstance()
-        ar_root = parent.createARPackage("AUTOSAR")
-        mock_element = ConcreteReferrable(ar_root, "TestElement")
-        obj.addElement(mock_element)
-
-        # Get element with default type=None (should return the element)
-        result = obj.getElement("TestElement")  # type defaults to None
-        assert result == mock_element
-
-    def test_ar_element_initialization(self):
-        """
-        Test ARElement initialization to cover line 372 in super().__init__ call.
-        """
-        parent = AUTOSAR.getInstance()
-        ar_root = parent.createARPackage("AUTOSAR")
-
-        class ConcreteARElement(ARElement):
-            def __init__(self, parent, short_name):
-                super().__init__(parent, short_name)
-
-        # This should trigger the super().__init__(parent, short_name) call in ARElement
-        obj = ConcreteARElement(ar_root, "TestARElement")
-        assert obj.getShortName() == "TestARElement"
-        assert obj.getParent() == ar_root
-
-    def test_describable_initialization(self):
-        """
-        Test Describable initialization to cover line 384 in super().__init__ call.
-        """
-
-        class ConcreteDescribable(Describable):
-            def __init__(self):
-                super().__init__()
-
-        # This should trigger the super().__init__() call in Describable
-        obj = ConcreteDescribable()
-        assert obj is not None
-
-    def test_get_set_category_identifiable_with_object(self):
-        """
-        Test setCategory in Identifiable class with non-string value to cover else branch.
-        """
-        parent = AUTOSAR.getInstance()
-        ar_root = parent.createARPackage("AUTOSAR")
-
-        class ConcreteIdentifiable(Identifiable):
-            def __init__(self, parent, short_name):
-                super().__init__(parent, short_name)
-
-        obj = ConcreteIdentifiable(ar_root, "TestName")
-
-        # Test with string value (if not already tested thoroughly)
-        obj.setCategory("TestCategory")
-        # The string case calls CategoryString().setValue(value)
-
-        # Test with object value (the else branch at line 372 - wait, that's not right)
-        # Actually for the else branch in setCategory method of Identifiable class
-        from armodel.models.M2.AUTOSARTemplates.GenericStructure.GeneralTemplateClasses.PrimitiveTypes import CategoryString
-
-        category_obj = CategoryString().setValue("ObjectCategory")
-        obj.setCategory(category_obj)  # This should go to the else branch
-        assert obj.getCategory() is category_obj
-
-    def test_is_element_exists(self):
-        """
-        Test IsElementExists method.
-        """
-
-        class ConcreteCollectableElement(CollectableElement):
-            def __init__(self):
-                super().__init__()
-
-        obj = ConcreteCollectableElement()
-
-        # Initially should return False
-        assert obj.IsElementExists("NonExistent") is False
-
-        # Add an element
-        class MockReferrable:
-            def __init__(self, short_name):
-                self.short_name = short_name
-
-            def getShortName(self):
-                return self.short_name
-
-        mock_element = MockReferrable("TestElement")
-        obj.addElement(mock_element)
-
-        # Should return True
-        assert obj.IsElementExists("TestElement") is True
-        assert obj.IsElementExists("NonExistent") is False
-
-    def test_remove_element(self):
-        """
-        Test removeElement method.
-        """
-
-        class ConcreteCollectableElement(CollectableElement):
-            def __init__(self):
-                super().__init__()
-
-        obj = ConcreteCollectableElement()
-
-        # Add an element - use a proper Referrable implementation
-        class ConcreteReferrable(Referrable):
-            def __init__(self, parent, short_name):
-                super().__init__(parent, short_name)
-
-        parent = AUTOSAR.getInstance()
-        ar_root = parent.createARPackage("AUTOSAR")
-        mock_element = ConcreteReferrable(ar_root, "TestElement")
-        obj.addElement(mock_element)
-
-        # Verify element exists
-        assert obj.IsElementExists("TestElement") is True
-        assert obj.getTotalElement() == 1
-
-        # Remove the element
-        obj.removeElement("TestElement")
-
-        # Note: There appears to be a bug in the source code where the key remains in element_mappings
-        # even after all elements are removed, so IsElementExists still returns True
-        # Let's just check that the total element count is 0
-        assert obj.getTotalElement() == 0
-
-
 class TestIdentifiable:
     """
     Test class for Identifiable functionality.
@@ -771,42 +389,6 @@ class TestIdentifiable:
         category_obj = CategoryString().setValue("NewCategory")
         obj.setCategory(category_obj)
         assert obj.getCategory() is category_obj
-
-
-class TestPackageableElement:
-    """
-    Test class for PackageableElement functionality.
-    """
-
-    def test_abstract_initialization(self):
-        """
-        Test that PackageableElement cannot be instantiated directly (abstract class).
-        """
-        parent = AUTOSAR.getInstance()
-        ar_root = parent.createARPackage("AUTOSAR")
-        try:
-            _obj = PackageableElement(ar_root, "TestPackageableElement")
-            assert False, "PackageableElement should not be instantiable"
-        except TypeError:
-            pass  # Expected behavior
-
-
-class TestARElement:
-    """
-    Test class for ARElement functionality.
-    """
-
-    def test_abstract_initialization(self):
-        """
-        Test that ARElement cannot be instantiated directly (abstract class).
-        """
-        parent = AUTOSAR.getInstance()
-        ar_root = parent.createARPackage("AUTOSAR")
-        try:
-            _obj = ARElement(ar_root, "TestARElement")
-            assert False, "ARElement should not be instantiable"
-        except TypeError:
-            pass  # Expected behavior
 
 
 class TestDescribable:
