+++ tests/test_armodel/models/M2/AUTOSARTemplates/GenericStructure/test_VariantHandling.py
@@ -6,12 +6,12 @@ from armodel.models.M2.AUTOSARTemplates.CommonStructure.StandardizationTemplate.
     BlueprintGenerator,
 )
 from armodel.models.M2.AUTOSARTemplates.GenericStructure.GeneralTemplateClasses.ARPackage import (
+    ARElement,
     ARPackage,
 )
 from armodel.models.M2.AUTOSARTemplates.GenericStructure.GeneralTemplateClasses.Enumerations import (
     BindingTimeEnum,
 )
-from armodel.models.M2.AUTOSARTemplates.GenericStructure.GeneralTemplateClasses.Identifiable import ARElement
 from armodel.models.M2.AUTOSARTemplates.GenericStructure.GeneralTemplateClasses.PrimitiveTypes import (
     ARNumerical,
     Identifier,
